Refactor: Atomic 패턴에서 /atoms를 제외한 모든 폴더 삭제, /common으로 통합
