+++ src/components/common/BottomTabNav.tsx
@@ -2,7 +2,7 @@
 
 import { usePathname, useRouter } from 'next/navigation';
 import { TABS } from '@/constants/BottomTabNav';
-import { BottomTabNavBtn } from '../molecules';
+import { BottomTabNavBtn } from '../common';
 
 type TabName = keyof typeof TABS;
 type RouteName = (typeof TABS)[TabName]['route'];
