+++ src/components/common/index.ts
@@ -1,4 +1,5 @@
 export { default as BottomTabNav } from './BottomTabNav';
+export { default as BottomTabNavBtn } from './BottomTabNavBtn';
 export { default as FixedBottomButton } from './FixedBottomButton';
 export { default as Header } from './Header';
 export { default as Modal } from './Modal';
